requirements.txt
streamlit
yfinance
pandas
pandas_ta
numpy
matplotlib
scikit-learn
tensorflow-cpu
plotly